Added the navbar for none mobile platforms
