+++ public/scss/components/_index.scss
@@ -1,5 +1,5 @@
 // Import all the components into this file 
-@use './navbar';
+@use './navigation';
 @use './header';
 @use './social_handles';
 @use './footer';
