--- public/scss/components/_navbar.scss
@@ -1,84 +1,0 @@
-@use '../custom' as *;
-@use '../utils/mixins' as *;
-.menu {
-    @include gradient;
-    .navbar-toggler {
-        border: 0 !important;
-        width: fit-content;
-        padding: 0 5px 0 0;
-        margin: 0;
-        &:focus {
-            border: 0 !important;
-            outline: 0 !important;
-            box-shadow: none;
-        }
-        &-icon {
-            background: none;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            // padding: 20px;
-            font-size: 25px;
-            i {
-                display: none;
-            }
-        }
-        &:not([aria-expanded="true"]) i {
-            &:first-of-type {
-                display: block;
-            }
-            &:last-of-type {
-                display: none;
-            }
-        }
-        &[aria-expanded="true"] {
-            i {
-                &:first-of-type {
-                    display: none;
-                }
-                &:last-of-type {
-                    display: block;
-                }
-            }
-        }
-    }
-    li {
-        padding: 0 0.7rem;
-        a {
-            color: $white;
-            padding: 0 0.7rem;
-            text-transform: capitalize;
-            font-weight: 600;
-        }
-        &:hover {
-            color: $white;
-        }
-    }
-    .nav__handles {
-        color: $white;
-        margin-top: 0;
-        font-size: 20px;
-        display: none;
-    }
-}
-
-@media (max-width:760px) {
-    .menu {
-        max-width: 100%;
-        text-align: center;
-        .navbar-collapse {
-            padding: 10px 0;
-            justify-content: flex-start !important;
-            ul li {
-                padding: 10px 0;
-            }
-        }
-        button {
-            margin-top: 10px;
-            display: inline-block;
-        }
-        .nav__handles {
-            display: flex;
-        }
-    }
-}
