+++ public/scss/components/_navigation.scss
@@ -1,0 +1,212 @@
+@use '../custom' as *;
+@use '../utils/mixins' as *;
+
+
+
+nav:not(.standard) {
+    position: fixed;
+    // width: 50px;
+    top: 10%;
+    right:5px;
+    background-color: $primary;
+    color: $white;
+    border-radius: 20px;
+
+    z-index: 5;
+
+    a {
+        // display: block;
+        position: relative;
+
+        @include transitionProperty;
+        
+        color: $white;
+        font-size: $medium;
+        padding: 10px;
+        
+        width:100%;
+
+        display: flex;
+        align-items: center;
+        justify-content: center;
+
+        background-color: $primary;
+
+        border-radius: 15px;
+
+       
+        
+        span {
+            display: none;
+            margin: 0;
+            font-weight: 400;
+            font-size: calc($small + 4px);
+        }
+    }
+
+    ul{
+        list-style: none;
+        margin: 0;
+        padding: 0;
+        width: 50px;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+        row-gap: 2px;
+        
+
+        li{
+            display: block;
+            padding:10px;
+            @include transitionProperty;
+
+            &:hover>a{
+                transform: translateX(-50%);
+                justify-content:flex-start;
+                width: 130% !important;
+                min-width: 120px;
+                i {
+                    margin: 0 10px;
+                }
+                &::after{
+                    content:attr(data-name);
+                    position: absolute;
+                    left: 55px;
+                    top: 10%;
+                    width: 100%;
+                }
+                
+            }
+            &:first-of-type:hover>a{
+                transform: translate(-48%, -22%);
+            }
+            &:last-of-type:hover>a{
+                transform: translate(-48%, 22%);
+            }
+            
+        }
+    }
+    
+}
+
+.menu {
+    @include gradient;
+
+    .navbar-toggler {
+        border: 0 !important;
+        width: fit-content;
+        padding: 0 5px 0 0;
+        margin: 0;
+
+        &:focus {
+            border: 0 !important;
+            outline: 0 !important;
+            box-shadow: none;
+        }
+
+        &-icon {
+            background: none;
+            display: flex;
+            align-items: center;
+            justify-content: center;
+            // padding: 20px;
+            font-size: 25px;
+
+            i {
+                display: none;
+            }
+        }
+
+        &:not([aria-expanded="true"]) i {
+            &:first-of-type {
+                display: block;
+            }
+
+            &:last-of-type {
+                display: none;
+            }
+        }
+
+        &[aria-expanded="true"] {
+            i {
+                &:first-of-type {
+                    display: none;
+                }
+
+                &:last-of-type {
+                    display: block;
+                }
+            }
+        }
+    }
+
+    li {
+        padding: 0.7rem;
+
+        a {
+            color: $white;
+            padding: 0 0.7rem;
+            text-transform: capitalize;
+            font-weight: 600;
+        }
+
+        &:hover {
+            color: $white;
+        }
+    }
+
+    .nav__handles {
+        color: $white;
+        margin-top: 0;
+        font-size: 20px;
+        display: none;
+        margin: 15px 0;
+    }
+}
+
+
+nav {
+    &.standard {
+        display: none;
+    }
+
+    &:not(.standard) {
+        display: block;
+    }
+}
+
+
+@media (max-width:760px) {
+    nav{
+        &.standard{
+            display: block;
+        }
+        &:not(.standard){
+            display: none;
+        }
+    }
+
+    .menu {
+        max-width: 100%;
+        text-align: center;
+
+        .navbar-collapse {
+            padding: 10px 0;
+            justify-content: flex-start !important;
+
+            ul li {
+                padding: 10px 0;
+            }
+        }
+
+        button {
+            margin-top: 10px;
+            display: inline-block;
+        }
+
+        .nav__handles {
+            display: flex;
+        }
+    }
+}
